+++ packages/root/types/types/index.d.ts
@@ -1,5 +1,5 @@
 export { PendingHistory, History } from "./hickory";
-export { LocationComponents, PartialLocation, SessionLocation, AnyLocation, Location } from "./location";
+export { LocationComponents, PartialLocation, SessionLocation, AnyLocation, RawLocation } from "./location";
 export { KeyFns } from "./keyGenerator";
 export { LocationUtilOptions, QueryFunctions, RawPathname, LocationUtils } from "./locationUtils";
 export { NavigationInfo, ConfirmationFunction, ConfirmationMethods, BlockingHistory } from "./navigationConfirmation";