+++ packages/root/types/types/location.d.ts
@@ -5,11 +5,11 @@ export interface LocationComponents {
     state?: any;
 }
 export declare type PartialLocation = Partial<LocationComponents>;
-export interface Location extends LocationComponents {
+export interface RawLocation extends LocationComponents {
     rawPathname: string;
 }
 export declare type Key = [number, number];
-export interface SessionLocation extends Location {
+export interface SessionLocation extends RawLocation {
     key: Key;
 }
 export declare type AnyLocation = SessionLocation | PartialLocation;